requirements.txt
Flask
gunicorn
flask-redis
numpy
scipy
Pillow
